requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Session==0.5.0
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.32.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==78.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3